feat: split auth into feature component and dialog feature
