+++ src/features/authentication/authentication-dialog.tsx
@@ -1,0 +1,47 @@
+'use client';
+
+import { usePathname, useRouter } from 'next/navigation';
+import { useEffect } from 'react';
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+} from '~/components/ui/dialog';
+import { useCurrentUserFromStore } from '~/hooks/use-current-user';
+import { useCreateQueryString } from '~/utils/create-query-string';
+import { AuthenticationFeature } from './authentication';
+import { useAuthenticationModal } from './hooks/use-authentication-dialog';
+
+export function AuthenticationDialogFeature() {
+  const router = useRouter();
+  const pathname = usePathname();
+  const user = useCurrentUserFromStore();
+  const createQueryString = useCreateQueryString();
+  const { isOpen, closeDialog } = useAuthenticationModal();
+
+  const handleOpenChange = (open: boolean) => {
+    if (open) return;
+    router.replace(`${pathname}/${createQueryString('action')}`);
+    closeDialog();
+  };
+
+  useEffect(() => {
+    if (!user) return;
+    closeDialog();
+  }, [user, closeDialog]);
+
+  return (
+    <Dialog open={isOpen} onOpenChange={handleOpenChange}>
+      <DialogContent className="border border-neutral-700 bg-neutral-800 sm:max-w-md">
+        <DialogHeader>
+          <DialogTitle>Welcome back!</DialogTitle>
+          <DialogDescription>Log in to your account</DialogDescription>
+        </DialogHeader>
+
+        <AuthenticationFeature />
+      </DialogContent>
+    </Dialog>
+  );
+}
